rviz: fixed bug #5233: large occupancy grid map crashes rviz.  Also changed default update interval for TF display from 0.0 to 0.05, to avoid the performance problems that are happening with the 0 interval.  This is not a proper fix, but should make it work better until I can get around to it.
git-svn-id: https://code.ros.org/svn/ros-pkg/stacks/visualization/trunk@38162 eb33c2ac-9c88-4c90-87e0-44a10359b0c3

+++ src/rviz/default_plugin/map_display.cpp
@@ -56,8 +56,8 @@ MapDisplay::MapDisplay( const std::string& name, VisualizationManager* manager )
 , manual_object_( NULL )
 , loaded_( false )
 , resolution_( 0.0f )
-, width_( 0.0f )
-, height_( 0.0f )
+, width_( 0 )
+, height_( 0 )
 , position_(Ogre::Vector3::ZERO)
 , orientation_(Ogre::Quaternion::IDENTITY)
 , draw_under_(false)
@@ -342,7 +342,7 @@ void MapDisplay::load(const nav_msgs::OccupancyGrid::ConstPtr& msg)
 
     ROS_WARN("Failed to create full-size map texture, likely because your graphics card does not support textures of size > 2048.  Downsampling to [%d x %d]...", (int)width, (int)height);
     //ROS_INFO("Stream size [%d], width [%f], height [%f], w * h [%f]", pixel_stream->size(), width_, height_, width_ * height_);
-    image.loadRawData(pixel_stream, (int)width_, (int)height_, Ogre::PF_L8);
+    image.loadRawData(pixel_stream, width_, height_, Ogre::PF_L8);
     image.resize(width, height, Ogre::Image::FILTER_NEAREST);
     ss << "Downsampled";
     texture_ = Ogre::TextureManager::getSingleton().loadImage(ss.str(), Ogre::ResourceGroupManager::DEFAULT_RESOURCE_GROUP_NAME, image);
@@ -477,11 +477,11 @@ void MapDisplay::createProperties()
   resolution_property_ = property_manager_->createProperty<FloatProperty>( "Resolution", property_prefix_, boost::bind( &MapDisplay::getResolution, this ),
                                                                             FloatProperty::Setter(), parent_category_, this );
   setPropertyHelpText(resolution_property_, "Resolution of the map. (not editable)");
-  width_property_ = property_manager_->createProperty<FloatProperty>( "Width", property_prefix_, boost::bind( &MapDisplay::getWidth, this ),
-                                                                       FloatProperty::Setter(), parent_category_, this );
+  width_property_ = property_manager_->createProperty<IntProperty>( "Width", property_prefix_, boost::bind( &MapDisplay::getWidth, this ),
+                                                                    IntProperty::Setter(), parent_category_, this );
   setPropertyHelpText(width_property_, "Width of the map, in meters. (not editable)");
-  height_property_ = property_manager_->createProperty<FloatProperty>( "Height", property_prefix_, boost::bind( &MapDisplay::getHeight, this ),
-                                                                        FloatProperty::Setter(), parent_category_, this );
+  height_property_ = property_manager_->createProperty<IntProperty>( "Height", property_prefix_, boost::bind( &MapDisplay::getHeight, this ),
+                                                                     IntProperty::Setter(), parent_category_, this );
   setPropertyHelpText(height_property_, "Height of the map, in meters. (not editable)");
   position_property_ = property_manager_->createProperty<Vector3Property>( "Position", property_prefix_, boost::bind( &MapDisplay::getPosition, this ),
                                                                            Vector3Property::Setter(), parent_category_, this );
